requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
redis
qdrant-client
pydantic
python-dotenv
sentence-transformers
pdfplumber
